Remove IRIS code except spectrograph reader.
