--- sunraster/instr/iris/data/README.rst
@@ -1,6 +1,0 @@
-Data directory
-==============
-
-This directory contains data files included with the affiliated package source
-code distribution. Note that this is intended only for relatively small files
-- large files should be externally hosted and downloaded as needed.
