--- sunraster/instr/iris/data/v40-table2000.csv
@@ -1,1 +1,0 @@
-,OBS ID,Size + description,Impact on cadence (>1 = slower),Impact on SG datarate,Impact on SJI datarate,0,C II   Si IV   Mg II h/k   Mg II w   ,1,1,1,100,C II   Si IV   Mg II h/k   Mg II w s   ,1,1,0.791666667,200,C II   Si IV   Mg II w s   ,1,1,0.541666667,300,C II   Mg II h/k   Mg II w s   ,1,1,0.541666667,400,Si IV   Mg II h/k   Mg II w s   ,1,1,0.541666667,500,C II   Mg II w s   ,1,1,0.291666667,600,Si IV   Mg II w s   ,1,1,0.291666667,700,Mg II h/k   Mg II w s   ,1,1,0.291666667,800,Si IV   Mg II h/k   Mg II w   ,1,1,0.75,900,C II   Mg II h/k   Mg II w   ,1,1,0.75,1000,C II   Si IV   Mg II w   ,1,1,0.75,1100,C II   Si IV   Mg II h/k   ,1,1,0.75,1200,C II   Si IV   ,1,1,0.5,1300,C II   Mg II h/k   ,1,1,0.5,1400,Si IV   Mg II h/k   ,1,1,0.5,1500,C II   ,1,1,0.25,1600,Si IV   ,1,1,0.25,1700,Mg II h/k   ,1,1,0.25,1800,Mg II w   ,1,1,0.25,1900,Mg II h/k   Mg II w   ,1,1,0.5,0,Exposure 1s,1,1,1,2000,Deep x 0.5,0.5,2,1,4000,Deep x 2,2,0.5,1,6000,Deep x 4,4,0.25,1,8000,Deep x 8,8,0.125,1,10000,Deep x 15,15,0.066666667,1,12000,Deep x 30,30,0.033333333,1,0,"Spatial x 1, Spectral x 1",1,1,1,20000,"Spatial x 1, Spectral x 2",1,0.5,1,40000,"Spatial x 1, Spectral x 4",1,0.25,1,60000,"Spatial x 1, Spectral x 8",1,0.125,1,80000,"Spatial x 2, Spectral x 1",1,0.5,1,100000,"Spatial x 2, Spectral x 2",1,0.25,1,120000,"Spatial x 2, Spectral x 4",1,0.125,1,140000,"Spatial x 2, Spectral x 8",1,0.0625,1,160000,"Spatial x 4, Spectral x 1",1,0.25,1,180000,"Spatial x 4, Spectral x 2",1,0.125,1,200000,"Spatial x 4, Spectral x 4",1,0.0625,1,220000,"Spatial x 4, Spectral x 8",1,0.03125,1,0,FUV binned same as NUV,1,1,1,250000,FUV spectrally rebinned x 2,1,0.5,1,500000,FUV spectrally rebinned x 4,1,0.25,1,750000,FUV spectrally rebinned x 8,1,0.125,1,0,SJI cadence 10s,1,1,1,1000000,SJI cadence 0.25x faster,1,1,0.25,2000000,SJI cadence 0.5x faster,1,1,0.5,3000000,SJI cadence 3x faster,1,1,2,4000000,SJI cadence 10x faster,1,1,10,8000000,AEC enabled,1,1.5,1.5,9000000,Flatfield and calibration,1,1,1,0,Lossy compression,1,1,1,10000000,Lossless compression,1,1.5,1.5,0,Large linelist,1,1,1,40000000,Small linelist,1,1,1,180000000,Full readout,1,1,1
